+++ RELEASE_NOTES.md
@@ -73,6 +73,7 @@
     * [SER-2745] - Fix file name in the bulk import downloads to use file name from the API
     * [SER-2847] - Allow overide of OCR checks if a user has the SKIP_VERIFICATION_CLIENT permission.
     * [SER-2892] - Search fields on Client Search and Filter Bars are labeled with "Search for Client" and "Filter by Client" respectively and fix null objects in the view jobs interface.
+    * [SER-2926] - Remove loan payment button from the UI
 
 ## Version 1.0.0 - for use with Fineract Web App